update color ramp

+++ GeometryShader/main_geometryshader.cpp
@@ -147,7 +147,7 @@ int boom()
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
         // configure transformation matrices
-        glm::mat4 projection = glm::perspective(camera.Fov, (float)SCR_WIDTH / (float)SCR_HEIGHT, 1.0f, 100.0f);
+        glm::mat4 projection = glm::perspective((float)camera.Fov, (float)SCR_WIDTH / (float)SCR_HEIGHT, 1.0f, 100.0f);
         glm::mat4 view = camera.getViewMatrix();;
         glm::mat4 model = glm::mat4(1.0f);
         shader.use();
@@ -220,7 +220,7 @@ int normal()
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
         // configure transformation matrices
-        glm::mat4 projection = glm::perspective(camera.Fov, (float)SCR_WIDTH / (float)SCR_HEIGHT, 1.0f, 100.0f);
+        glm::mat4 projection = glm::perspective((float)camera.Fov, (float)SCR_WIDTH / (float)SCR_HEIGHT, 1.0f, 100.0f);
         glm::mat4 view = camera.getViewMatrix();;
         glm::mat4 model = glm::mat4(1.0f);
         shader.use();
